lots of db work comlpeted
So, when the user logs in, any ratings they have made will be stored in the user object so the page can find the recipeid in that list of ratings and put it on the page. The user object has a method to get these ratings by id.

The recipe object has a dictionary containing in it the average rating and count of votes by type of rating in it. This will be input into the recipe object when it loads it from the api. So, only when shown on the homepage or when it is in a search results list. When a recipe is clicked to be viewed, it will look in the database for the recipe id, and pull info from there related to the recipe. (I haven't set this up yet but will do so next), And if it isn't in the db, it will pull the info from the api. If the user has not rated the object a tuple<taste, difficulty> of 0, 0 is returned when the page looks for the user ratings in the user object. If the user is not logged in, I am not sure what to display as the user object will be blank.